feat(core): add archive and path helpers
